@@ -1,8 +1,8 @@
 [kivy]
 cython_min=0.29.1
-cython_max=3.0.0
+cython_max=3.0.11
 cython_exclude=
-python_versions=3.7 - 3.11
+python_versions=3.8 - 3.13
 
 [coverage:run]
 parallel = True
@@ -14,13 +14,15 @@ plugins =
 concurrency = thread, multiprocessing
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires =
     Kivy-Garden>=0.1.4
     docutils
     pygments
-    kivy_deps.angle~=0.3.3; sys_platform == "win32"
-    kivy_deps.sdl2~=0.6.0; sys_platform == "win32"
+    requests
+    filetype
+    kivy_deps.angle~=0.4.0; sys_platform == "win32"
+    kivy_deps.sdl2~=0.8.0; sys_platform == "win32"
     kivy_deps.glew~=0.3.1; sys_platform == "win32"
     pypiwin32; sys_platform == "win32"
 dependency_links = https://github.com/kivy-garden/garden/archive/master.zip
@@ -34,48 +36,30 @@ dev =
     pytest-timeout
     pytest-benchmark
     pyinstaller
-    sphinx<=6.2.1
-    sphinxcontrib-blockdiag
-    sphinxcontrib-seqdiag
-    sphinxcontrib-actdiag
-    sphinxcontrib-nwdiag
-    sphinxcontrib-jquery
-    funcparserlib==1.0.0a0
+    sphinx~=6.2.1
+    sphinxcontrib-jquery~=4.1
     kivy_deps.gstreamer_dev~=0.3.3; sys_platform == "win32"
-    kivy_deps.sdl2_dev~=0.6.0; sys_platform == "win32"
+    kivy_deps.sdl2_dev~=0.8.0; sys_platform == "win32"
     kivy_deps.glew_dev~=0.3.1; sys_platform == "win32"
     flake8
     pre-commit
     responses
 base =
-    pillow==9.5.0
-    requests
-    docutils
-    pygments
-    kivy_deps.angle~=0.3.3; sys_platform == "win32"
-    kivy_deps.sdl2~=0.6.0; sys_platform == "win32"
-    kivy_deps.glew~=0.3.1; sys_platform == "win32"
-    pypiwin32; sys_platform == "win32"
+    pillow>=9.5.0,<11
 media =
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
     ffpyplayer; sys_platform == "linux" or sys_platform == "darwin"
 full =
-    pillow==9.5.0
-    docutils
-    pygments
+    pillow>=9.5.0,<11
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
-    kivy_deps.angle~=0.3.3; sys_platform == "win32"
-    kivy_deps.sdl2~=0.6.0; sys_platform == "win32"
-    kivy_deps.glew~=0.3.1; sys_platform == "win32"
     ffpyplayer; sys_platform == "linux" or sys_platform == "darwin"
-    pypiwin32; sys_platform == "win32"
 gstreamer =
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
     # don't use 0.4.0 because it was deleted
 angle =
-    kivy_deps.angle~=0.3.3; sys_platform == "win32"
+    kivy_deps.angle~=0.4.0; sys_platform == "win32"
 sdl2 =
-    kivy_deps.sdl2~=0.6.0; sys_platform == "win32"
+    kivy_deps.sdl2~=0.8.0; sys_platform == "win32"
 glew =
     kivy_deps.glew~=0.3.1; sys_platform == "win32"
 
